+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 [Semantic Versioning](http://semver.org/).
 
 ## 1.17.4.dev (development stage/unreleased/unstable)
+
+### Added
+- binance.com testnets
+
 ## 1.17.4
 ### Added
 - `replace_stream()`: new_stream_label=None, new_stream_buffer_name=False, new_symbol=False, new_api_key=False,